+++ app/scripts/data_init.ts
@@ -82,7 +82,7 @@ function transformConfig(config: Config, infos: InfoData): ConfigTransformed {
         const [custom_attributes, _custom_attributes_order] = transformAttributes2<CustomAttribute>("custom_attributes")
 
         return {
-            ...omit(corpus, "pos_attributes"),
+            ...omit(corpus, ["pos_attributes", "limited_access"]),
             attributes,
             struct_attributes,
             custom_attributes,
@@ -93,6 +93,7 @@ function transformConfig(config: Config, infos: InfoData): ConfigTransformed {
             within: contextWithinFix(corpus["within"]),
             info: infos[corpus.id].info,
             private_struct_attributes: infos[corpus.id].private_struct_attributes,
+            protected: infos[corpus.id].info["Protected"] === "true",
         }
     }