+++ app/scripts/parallel/corpus-set-parallel.ts
@@ -25,10 +25,16 @@ export class CorpusSetParallel extends CorpusSet {
     }
 
     pickFrom(source: CorpusSetParallel, ids: string[]): void {
-        // Include hidden linked corpora
-        const corpora = ids.flatMap((id) => source.getLinked(source.get(id)))
-        ids = corpora.map((corpus) => corpus.id)
-        super.pickFrom(source, ids)
+        // Include linked corpora, except if linked from pivot corpus
+        const corpora = ids
+            .flatMap((id) => id.split("|"))
+            .flatMap((id) => {
+                const corpus = source.get(id)
+                const isPivot = corpus.linked_to.length > 1
+                return isPivot ? corpus : source.getLinked(corpus)
+            })
+        const idsAll = corpora.map((corpus) => corpus.id)
+        super.pickFrom(source, idsAll)
     }
 
     setActiveLangs(langlist: string[]): void {
@@ -108,16 +114,14 @@ export class CorpusSetParallel extends CorpusSet {
     getAttributeQuery(attr: "context" | "within"): string {
         const struct = this.getLinksFromLangs(this.activeLangs)
         const output: string[][] = struct.map((corps) => {
-            const mainId = corps[0].id.toUpperCase()
-            const mainIsPivot = !!corps[0].pivot
-
-            const other = corps.slice(1)
-
-            const pair = other.map(function (corp) {
-                const a = mainIsPivot ? Object.keys(corp[attr])[0] : Object.keys(corps[0][attr])[0]
-                return mainId + "|" + corp.id.toUpperCase() + ":" + a
+            const [main, ...others] = corps
+            const isPivot = main.linked_to.length > 1
+            return others.map(function (other) {
+                // For pivot corpus, use the linked corpus config instead
+                const corpus = isPivot ? other : main
+                const value = Object.keys(corpus[attr])[0]
+                return `${main.id}|${other.id}`.toUpperCase() + ":" + value
             })
-            return pair
         })
 
         return output.join(",")
@@ -135,21 +139,19 @@ export class CorpusSetParallel extends CorpusSet {
     }
 
     stringify(onlyMain?: boolean): string {
-        let struct = this.getLinksFromLangs(this.activeLangs)
-        if (onlyMain) {
-            struct = struct.map((pair) => {
-                return pair.filter((item) => {
-                    return item.lang === this.activeLangs[0]
-                })
-            })
+        const lists = this.getLinksFromLangs(this.activeLangs)
 
-            return struct
-                .flat()
-                .map((corpus) => corpus.id.toUpperCase())
-                .join()
+        if (onlyMain) {
+            // Select corpora in the first search language
+            const corpora = lists.flat().filter((item) => item.lang === this.activeLangs[0])
+            return corpora.map((corpus) => corpus.id.toUpperCase()).join()
         }
 
-        return struct.map(([a, b]) => `${a.id}|${b.id}`.toUpperCase()).join()
+        // Format pairs like X-SV|X-DA,X-SV|X-EN...
+        return lists
+            .flatMap(([main, ...others]) => others.map((other) => `${main.id}|${other.id}`))
+            .join()
+            .toUpperCase()
     }
 
     get(corpusID: string): CorpusTransformed<CorpusParallel> {